Move Hypermedia decorator path from infrastructure to application
