+++ src/common/base/application/decorator/hypermedia.decorator.ts
